完成报警接口,接受请求,获取汽车信息(缓存)和规则信息,解析规则用js计算公式,用比较器判断是否符合条件,日志记录

+++ src/main/java/com/xiaomi/domain/rule/Rate.java
@@ -11,7 +11,7 @@ import java.util.List;
 public class Rate {
 
     @ApiModelProperty(value = "告警级别")
-    private Integer warnLever;
+    private Byte warnLever;
 
     @ApiModelProperty(value = "条件列表")
     private List<Condition> condition;
